Refactored code out of init files
