+++ bandwidth/voice/api_exception_module.py
@@ -1,0 +1,27 @@
+class BandwidthVoiceAPIException(Exception):
+
+    """
+    Catapult API request exception
+    """
+
+    def __init__(self, status_code, message, **kwargs):
+        """
+        Initialize the catapult exception.
+        :type status_code: str
+        :param status_code: http status code
+        :type message: str
+        :param message: error message
+        :type code: str
+        :param code: optional error code
+
+        :rtype: bandwidth.catapult.CatapultException
+        :returns: instance of exception
+        """
+        self.status_code = status_code
+        self.message = message
+        self.code = kwargs.get('code')
+        if self.code is None:
+            self.code = str(status_code)
+
+    def __str__(self):
+        return 'Error %s: %s' % (self.code, self.message)
